This is the first version of the Picker Xaml control authored in cppwinrt
The Control is authored in cppwinrt and the the accompanying app XAMLHostingDesktop is the app that is used to host the control in a Win32 app. This provides the flexibility of using a Xaml file picker control that provides the modern look and feel while providing the ability to use the control in a Win32 application.